fix(audit): keep test quality detector focused
